pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Prowler Team"]
description = ""
name = "prowler-studio"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
bandit = "^1.7.10"
llama-index-core = "^0.12.0"
llama-index-embeddings-gemini = "^0.3.0"
llama-index-llms-gemini = "^0.4.0"
llama-index-llms-llama-cpp = "^0.3.0"
llama-index-readers-github = "^0.5.0"
pylint = "^3.3.1"
python = "^3.12"
safety = "^3.2.10"
vulture = "^2.13"
